completed ex

+++ src/shapes/Quadrilateral.java
@@ -1,23 +1,23 @@
 package shapes;
 
 public abstract class Quadrilateral extends Shape implements Measurable {
-    protected int length;
-    protected int width;
+    protected double length;
+    protected double width;
 
-    public Quadrilateral(int length, int width){
+    public Quadrilateral(double length, double width){
         this.length = length;
         this.width = width;
     }
 
-    public int getLength(){
+    public double getLength(){
         return length;
     }
 
-    public int getWidth() {
+    public double getWidth() {
         return width;
     }
 
-    public abstract void setWidth(int width);
+    public abstract void setWidth(double width);
 
-    public abstract void setLength(int length);
+    public abstract void setLength(double length);
 }
